Refactor thread components and update file naming convention
