+++ app/components/message.tsx
@@ -1,6 +1,6 @@
 import { useMutation, useQuery } from "convex/react"
 import dayjs from "dayjs"
-import { ChevronDownIcon, Edit2Icon, SmileIcon, SmilePlusIcon, TrashIcon } from "lucide-react"
+import { ChevronDownIcon, Edit2Icon, MessageSquareIcon, SmileIcon, SmilePlusIcon, TrashIcon } from "lucide-react"
 import { useEffect, useState } from "react"
 import { EmojiPicker, EmojiPickerContent, EmojiPickerFooter, EmojiPickerSearch } from "@/components/ui/emoji-picker"
 import { api } from "@/convex/_generated/api"
@@ -8,6 +8,7 @@ import type { Id } from "@/convex/_generated/dataModel"
 import { cn } from "@/lib/utils"
 import { ExpandableTextarea } from "./expandable-textarea"
 import { FilePill } from "./file-pill"
+import { ThreadIndicator } from "./thread/thread-indicator"
 import { Avatar, AvatarFallback, AvatarImage } from "./ui/avatar"
 import { Button } from "./ui/button"
 import { Popover, PopoverContent, PopoverTrigger } from "./ui/popover"
