+++ app/components/thread/thread-indicator.tsx
