+++ app/components/thread/thread-message-input.tsx
